[pkg/stanza/fileconsumer] Pull file attribute resolver out of reader
