+++ pkg/stanza/fileconsumer/reader.go
@@ -18,40 +18,11 @@ import (
 	"context"
 	"fmt"
 	"os"
-	"path/filepath"
 
 	"go.uber.org/zap"
 
 	"github.com/open-telemetry/opentelemetry-collector-contrib/pkg/stanza/operator/helper"
 )
-
-type FileAttributes struct {
-	Name         string
-	Path         string
-	NameResolved string
-	PathResolved string
-}
-
-// resolveFileAttributes resolves file attributes
-// and sets it to empty string in case of error
-func (f *Input) resolveFileAttributes(path string) *FileAttributes {
-	resolved, err := filepath.EvalSymlinks(path)
-	if err != nil {
-		f.Error(err)
-	}
-
-	abs, err := filepath.Abs(resolved)
-	if err != nil {
-		f.Error(err)
-	}
-
-	return &FileAttributes{
-		Path:         path,
-		Name:         filepath.Base(path),
-		PathResolved: abs,
-		NameResolved: filepath.Base(abs),
-	}
-}
 
 // Reader manages a single file
 type Reader struct {
@@ -71,12 +42,16 @@ type Reader struct {
 // NewReader creates a new file reader
 func (f *Input) NewReader(file *os.File, fp *Fingerprint, splitter *helper.Splitter) (*Reader, error) {
 	path := file.Name()
+	attrs, err := resolveFileAttributes(path)
+	if err != nil {
+		f.Errorf("resolve attributes: %w", err)
+	}
 	r := &Reader{
 		SugaredLogger:  f.SugaredLogger.With("path", path),
 		Fingerprint:    fp,
 		file:           file,
 		fileInput:      f,
-		fileAttributes: f.resolveFileAttributes(path),
+		fileAttributes: attrs,
 		splitter:       splitter,
 	}
 	return r, nil
